+++ src/FilteredActionList/FilteredActionList.tsx
@@ -1,13 +1,18 @@
-import React, {KeyboardEventHandler, useCallback, useMemo, useRef} from 'react'
+import React, {KeyboardEventHandler, useCallback, useEffect, useMemo, useRef} from 'react'
 import {GroupedListProps, ListPropsBase} from '../ActionList/List'
 import TextInput, {TextInputProps} from '../TextInput'
 import Box from '../Box'
+import Flex from '../Flex'
 import {ActionList} from '../ActionList'
 import Spinner from '../Spinner'
 import {useFocusZone} from '../hooks/useFocusZone'
 import {uniqueId} from '../utils/uniqueId'
 import {itemActiveDescendantClass} from '../ActionList/Item'
 import {useProvidedStateOrCreate} from '../hooks/useProvidedStateOrCreate'
+import styled from 'styled-components'
+import {get} from '../constants'
+import {useProvidedRefOrCreate} from '../hooks/useProvidedRefOrCreate'
+import useScrollFlash from '../hooks/useScrollFlash'
 
 export interface FilteredActionListProps extends Partial<Omit<GroupedListProps, keyof ListPropsBase>>, ListPropsBase {
   loading?: boolean
@@ -15,7 +20,36 @@ export interface FilteredActionListProps extends Partial<Omit<GroupedListProps, 
   filterValue?: string
   onFilterChange: (value: string, e: React.ChangeEvent<HTMLInputElement>) => void
   textInputProps?: Partial<Omit<TextInputProps, 'onChange'>>
+  inputRef?: React.RefObject<HTMLInputElement>
 }
+
+function scrollIntoViewingArea(
+  child: HTMLElement,
+  container: HTMLElement,
+  margin = 8,
+  behavior: ScrollBehavior = 'smooth'
+) {
+  const {top: childTop, bottom: childBottom} = child.getBoundingClientRect()
+  const {top: containerTop, bottom: containerBottom} = container.getBoundingClientRect()
+
+  const isChildTopAboveViewingArea = childTop < containerTop + margin
+  const isChildBottomBelowViewingArea = childBottom > containerBottom - margin
+
+  if (isChildTopAboveViewingArea) {
+    const scrollHeightToChildTop = childTop - containerTop + container.scrollTop
+    container.scrollTo({behavior, top: scrollHeightToChildTop - margin})
+  } else if (isChildBottomBelowViewingArea) {
+    const scrollHeightToChildBottom = childBottom - containerBottom + container.scrollTop
+    container.scrollTo({behavior, top: scrollHeightToChildBottom + margin})
+  }
+
+  // either completely in view or outside viewing area on both ends, don't scroll
+}
+
+const StyledHeader = styled.div`
+  box-shadow: 0 1px 0 ${get('colors.border.primary')};
+  z-index: 1;
+`
 
 export function FilteredActionList({
   loading = false,
@@ -24,6 +58,7 @@ export function FilteredActionList({
   onFilterChange,
   items,
   textInputProps,
+  inputRef: providedInputRef,
   ...listProps
 }: FilteredActionListProps): JSX.Element {
   const [filterValue, setInternalFilterValue] = useProvidedStateOrCreate(externalFilterValue, undefined, '')
@@ -36,8 +71,8 @@ export function FilteredActionList({
     [onFilterChange, setInternalFilterValue]
   )
 
-  const containerRef = useRef<HTMLInputElement>(null)
-  const inputRef = useRef<HTMLInputElement>(null)
+  const listContainerRef = useRef<HTMLDivElement>(null)
+  const inputRef = useProvidedRefOrCreate<HTMLInputElement>(providedInputRef)
   const activeDescendantRef = useRef<HTMLElement>()
   const listId = useMemo(uniqueId, [])
   const onInputKeyPress: KeyboardEventHandler = useCallback(
@@ -55,11 +90,11 @@ export function FilteredActionList({
   )
 
   useFocusZone({
-    containerRef,
+    containerRef: listContainerRef,
     focusOutBehavior: 'wrap',
     focusableElementFilter: element => {
       if (element instanceof HTMLInputElement) {
-        // No active-descendant focus on checkboxes in list items or filter input
+        // No active-descendant focus on checkboxes in list items
         return false
       }
       return true
@@ -74,26 +109,41 @@ export function FilteredActionList({
 
       if (current) {
         current.classList.add(itemActiveDescendantClass)
+
+        if (listContainerRef.current) {
+          scrollIntoViewingArea(current, listContainerRef.current)
+        }
       }
     }
   })
 
+  useEffect(() => {
+    // if items changed, we want to instantly move active descendant into view
+    if (activeDescendantRef.current && listContainerRef.current) {
+      scrollIntoViewingArea(activeDescendantRef.current, listContainerRef.current, undefined, 'auto')
+    }
+  }, [items])
+
+  useScrollFlash(listContainerRef)
+
   return (
-    <Box ref={containerRef} flexGrow={1} flexDirection="column">
-      <TextInput
-        ref={inputRef}
-        block
-        width="auto"
-        color="text.primary"
-        value={filterValue}
-        onChange={onInputChange}
-        onKeyPress={onInputKeyPress}
-        placeholder={placeholderText}
-        aria-label={placeholderText}
-        aria-controls={listId}
-        {...textInputProps}
-      />
-      <Box flexGrow={1} overflow="auto">
+    <Flex flexDirection="column" overflow="hidden">
+      <StyledHeader>
+        <TextInput
+          ref={inputRef}
+          block
+          width="auto"
+          color="text.primary"
+          value={filterValue}
+          onChange={onInputChange}
+          onKeyPress={onInputKeyPress}
+          placeholder={placeholderText}
+          aria-label={placeholderText}
+          aria-controls={listId}
+          {...textInputProps}
+        />
+      </StyledHeader>
+      <Box ref={listContainerRef} overflow="auto">
         {loading ? (
           <Box width="100%" display="flex" flexDirection="row" justifyContent="center" pt={6} pb={7}>
             <Spinner />
@@ -102,7 +152,7 @@ export function FilteredActionList({
           <ActionList items={items} {...listProps} role="listbox" id={listId} />
         )}
       </Box>
-    </Box>
+    </Flex>
   )
 }