+++ src/TextInput.tsx
@@ -59,7 +59,7 @@ const Wrapper = styled.span<StyledWrapperProps>`
   vertical-align: middle;
   background-repeat: no-repeat; // Repeat and position set for form states (success, error, etc)
   background-position: right 8px center; // For form validation. This keeps images 8px from right and centered vertically.
-  border: 1px solid ${get('colors.input.border')};
+  border: 1px solid ${get('colors.border.primary')};
   border-radius: ${get('radii.2')};
   outline: none;
   box-shadow: ${get('shadows.shadow.inset')};